tests/wpt/web-platform-tests/storage/storagemanager-persisted.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>StorageManager: persist()</title>
<meta name="help" href="https://storage.spec.whatwg.org/#dom-storagemanager-persist">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(function() {
    var promise = navigator.storage.persisted();
    assert_true(promise instanceof Promise,
        'navigator.storage.persisted() returned a Promise.');
    return promise.then(function (result) {
        assert_equals(typeof result, 'boolean', result + ' should be boolean');
    });
}, 'navigator.storage.persisted() returns a promise that resolves.');

</script>
